Fix object sync percentage calculation integer wraparound.
git-svn-id: https://svn.terracotta.org/repo/tc/dso/trunk@19094 7fc7bbf3-cf45-46d4-be06-341739edd864

+++ dso-l2/src/main/java/com/tc/l2/handler/L2ObjectSyncSendHandler.java
@@ -184,8 +184,8 @@ public class L2ObjectSyncSendHandler extends AbstractEventHandler {
       final int last = mosc.getTotalObjectsSynced();
       final int current = last + mosc.getSynchedOids().size();
       final int totalObjectsToSync = mosc.getTotalObjectsToSync();
-      final int lastPercent = (last * 100) / totalObjectsToSync;
-      final int currentPercent = (current * 100) / totalObjectsToSync;
+      final int lastPercent = (int) ((last * 100L) / totalObjectsToSync);
+      final int currentPercent = (int) ((current * 100L) / totalObjectsToSync);
 
       if (currentPercent > lastPercent) {
         logger.info("Sent " + current + " (" + currentPercent + "%) objects out of " + mosc.getTotalObjectsToSync()
